Split app setup from homepage
