+++ enacts/onset/maproom.py
@@ -23,7 +23,7 @@ from shapely import wkb
 from shapely.geometry.multipolygon import MultiPolygon
 import datetime
 
-from flask_app import FLASK
+from globals_ import FLASK
 
 GLOBAL_CONFIG = pingrid.load_config(os.environ["CONFIG"])
 CONFIG = GLOBAL_CONFIG["onset"]
